Replace use of nightly features to allow building againt stable rust
Features affected:
 - Create no longer keeps a global shared statsd client.
 - pcg32 seeding uses plain `time::precise_time_ns()` instead of assembly accessing register
 - No longer passing 'branch likely skipped' hint to the compiler
 - libc sendfile(fd) replaced by std rust UdpSocket::send

Likely impacts:

Performance may decrease slightly but is not expected to be noticeable. In the absence of hard performance data, it is hard to justify the use of the removed features. If performance is revealed to be an issue, other methods may be tried as documented in the updated `README.md` file.

Ergonomics of the statsd! macro are suboptimal and could rely on thread locals for more than just rng. This should be the object of a future PR.